general function exported to Utils classs
